+++ python/fate/components/entrypoint/cli/component/artifact_type_cli.py
