+++ python/fate/components/entrypoint/cli/component/desc_cli.py
